Add JSON API documentation
