+++ src/pages/api-graphql/index.css
@@ -1,0 +1,8 @@
+.api-docs-page-wrapper {
+    margin: auto;
+    max-width: 900px;
+}
+
+.section-text-wrapper {
+    margin: 10px 0;
+}
